qweqweqwe

+++ fixpoint_cpp_simulator/fixpoint_cpp_simulator/utils.h
@@ -4,8 +4,8 @@
 
 struct complex
 {
-	short real;
-	short imag;
+	int real;
+	int imag;
 };
 
 struct param
